requirements.txt
discord.py==2.5.1
python-dotenv==1.0.0
aiohttp==3.8.1
requests==2.26.0
beautifulsoup4==4.10.0